[YW2-18] comment 반영 및 refactoring
