+++ src/domain/user/src/main/kotlin/yapp.be.domain/port/inbound/model/ModifyUserTokenCommand.kt
@@ -1,8 +1,8 @@
 package yapp.be.domain.port.inbound.model
 
-import yapp.be.domain.model.OAuth2SecurityToken
 
 data class ModifyUserTokenCommand(
     val userId: Long,
-    val oAuth2SecurityToken: OAuth2SecurityToken
+    val oAuth2AccessToken: String,
+    val oAuth2RefreshToken: String,
 )
