+++ src/infrastructure/storage/build.gradle.kts
@@ -17,8 +17,6 @@ dependencies {
     implementation(project(":shelter"))
     implementation(project(":volunteerEvent"))
 
-    implementation("org.springframework.security:spring-security-oauth2-client")
-    implementation("org.springframework.boot:spring-boot-starter-security")
     implementation("org.springframework.boot:spring-boot-starter-data-jpa")
     implementation("com.querydsl:querydsl-jpa:$queryDslVersion")
